@GitBariAl e Io abbiamo iniziato a lavorare all'algoritmo di "AI"
(dobbiamo ancora finire)

+++ functions/initialization.c
@@ -5,7 +5,7 @@
 int** alloc_player_deck_memory (int src_size) {
     int **generatedMalloc;
     generatedMalloc = (int**) malloc(sizeof(int*) * src_size);      // allochiamo la memoria per x int* (che contengono i puntatori agli interi dell'array)
-    for (int i = 0; i < src_size; i ++){
+    for (size_t i = 0; i < src_size; i ++){
         generatedMalloc[i] = (int*) malloc(2 * sizeof(int));
     }
     return generatedMalloc;
